+++ Properties/AssemblyInfo.cs
@@ -1,0 +1,7 @@
+using System.Reflection;
+
+[assembly: AssemblyTitle("SourceTX Companion")]
+[assembly: AssemblyProduct("SourceTX Companion")]
+[assembly: AssemblyCompany("SourceTX Project")]
+[assembly: AssemblyVersion("0.2.0.0")]
+[assembly: AssemblyFileVersion("0.2.0.0")]